sycl: fix 9 code review issues — valid() semantics, profiling queues, pin guard, int64 safety
Fixes from comprehensive code review of the 11-commit coalescing series:

P0: Remove tracked binary executables, add tests/.gitignore
P1: Fix mem_handle::valid() — default-constructed handles no longer report
    valid (enables lazy handle population in get_data_ptr)
P1: Add enable_profiling to all directly-created SYCL queues via
    default_queue_properties() helper (5 queues were missing counter-based events)
P1: Remove const_cast in get_data_ptr — void* extra is directly accessible
P2: Add memoization (thread_local visited set) to
    tensor_depends_on_planned_host_weight to avoid exponential DAG traversal
P3: Unify pin/unpin skip guard into ggml_sycl_should_skip_pin_unpin() helper
P3: Document GGML_SYCL_ONEDNN_MUL env var in CLAUDE.md
P3: Use int64_t for total_blocks in SOA dequant kernel (overflow safety)
P3: Add STABLE_BASE get_caps + assertion in buffer_reset early return

+++ ggml/src/ggml-sycl/dequantize.hpp
@@ -369,8 +369,8 @@ template<typename dst_t>
 static void dequantize_block_q4_0_soa_rowmajor(const void * __restrict__ vx, dst_t * __restrict__ yy,
                                                 const int blocks_per_row, const int nrows,
                                                 const sycl::nd_item<3> & item) {
-    const int global_id  = item.get_global_id(2);
-    const int total_blocks = nrows * blocks_per_row;
+    const int     global_id    = item.get_global_id(2);
+    const int64_t total_blocks = (int64_t) nrows * blocks_per_row;
 
     if (global_id >= total_blocks) {
         return;
@@ -383,7 +383,7 @@ static void dequantize_block_q4_0_soa_rowmajor(const void * __restrict__ vx, dst
     const auto * qs = (const uint8_t *) vx + (int64_t) global_id * (QK4_0 / 2);
 
     // SOA: scales after ALL qs bytes
-    const int64_t total_qs_bytes = (int64_t) total_blocks * (QK4_0 / 2);
+    const int64_t total_qs_bytes = total_blocks * (QK4_0 / 2);
     const auto *  s_ptr = (const sycl::half *) ((const uint8_t *) vx + total_qs_bytes) + global_id;
     const float   d     = float(*s_ptr);
